Revert "fisher: store state in managed object"
This reverts commit e70e940027c04e019654c151324d64a7279232bb.

I've never cursed in a commit before, but this is the first time. Fuck
that gamepass version. I wasted an entire day on this version, and
honestly, it's the dumbest piece of *garbage* I've ever dealt with.
Every defined ECS system just doesn't fully run in that trash can.
And spending a whole day staring at that Goddamn Fucking Smart
capitalized path name and stupid backslash separator makes me feel
like I'm drowning in a sea of revulsion.